(#SIP 8) Tolerate vice
Implements #19 . 

`exploreDrains` will jump into a random location in the ShitDAO contract code.

The random jump is wrapped in a `delegatecall` context to swallow up any errors produced. This should make it extremely difficult to debug the function, as drains are full of bugs and the user will be outnumbered.

Tests: calling this function crashed remix and triggered a blue screen of death.